+++ src/components/social-post-modal/settings-section.tsx
@@ -1,9 +1,11 @@
+import { Panel } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
 import { store as editorStore } from '@wordpress/editor';
 import { __, _x } from '@wordpress/i18n';
 import usePublicizeConfig from '../../hooks/use-publicize-config';
 import { SharePostForm } from '../form/share-post-form';
 import { SharePostButton } from '../share-post';
+import { ScheduledShares } from './scheduled-shares';
 import styles from './styles.module.scss';
 
 /**
@@ -15,6 +17,7 @@ import styles from './styles.module.scss';
  */
 export function SettingsSection( { onReShared } ) {
 	const isPostPublished = useSelect( select => select( editorStore ).isCurrentPostPublished(), [] );
+	const postId = useSelect( select => select( editorStore ).getCurrentPostId(), [] );
 	const { isRePublicizeUpgradableViaUpsell } = usePublicizeConfig();
 
 	return (
@@ -40,6 +43,13 @@ export function SettingsSection( { onReShared } ) {
 					</div>
 				) }
 			</div>
+			{ isPostPublished ? (
+				<div>
+					<Panel>
+						<ScheduledShares postId={ postId } />
+					</Panel>
+				</div>
+			) : null }
 		</div>
 	);
 }